list implementations of atraverse
